### 2020-04-21 #30 @hhentschke
[Re] Hippocampal Phase-Amplitude Coupling unearthed again
This is a submission to the Ten Years Reproducibility Challenge.

**Original article:** Hentschke, H., Perkins, M.G., Pearce, R.A., & Banks, M.I. (2007) Muscarinic blockade weakens interaction of gamma with theta rhythms in mouse hippocampus. Eur.J.Neurosci., 26, 1642–1656.

**Scientific domain:**  Neuroscience
**Programming language:**  Matlab
**Suggested editor:**  Thomas Arildsen
