### 2020-05-07 #30 @TomDonoghue
Hey @oliviaguest - thanks for thinking of me!

The 10 years reproducibility challenge looks really interesting, and I would be happy to review this submission! Are there are specific guidelines for reviewing this kind of article / submission that I should be aware of?
