### 2020-05-24 #30 @TomDonoghue
Okay, so I have read through the original paper, the submitted ReScience paper, and looked through the repository. This paper reports on an attempt to re-run the analyses from a paper from 2007 examining theta & gamma oscillations, and cross frequency interactions, in mouse hippocampus, while manipulating muscarinic acetylcholinergic activity with atropine. I think this is a very interesting and overall good submission for the 10 years reproducibility challenge. I commend the author on picking up this very interesting challenge, and doing the work to step through old code, organize it, and try to get it all working! I have some comments, mainly addressed at hopefully trying to clarify the narrative of the paper, and a couple comments / questions about the organization of the repository.

### Review Related Comments

I have a couple open questions about the review process, that perhaps @oliviaguest can help with a bit. I read the general ReScience review comments, and the notes on this challenge in particular, but was left with a couple open questions.

- Am I supposed to be able to and try to re-run the code? The general ReScience guidelines suggest this, but I'm unclear whether this is also the case for this particular challenge. Since there is no data provided, I cannot run the code (I'm not sure whether this data could be released by the author). The descriptions in the repo are organized in terms of trying to get the code to run, but this isn't something I can currently try and do.
- To what extend am I supposed to review the code? Since I can't try and run it, and I'm unclear on if, and in what way I should review it. So far I haven't worked through the code too systematically.
- Are there guidelines for how much needs to be reproduced? This submission directly reproduces 4 sub-panels (5B, 5C, 7A, 9A) from the original paper. I'm not sure if there is a recommended threshold. (I don't necessarily think the paper should require more, but I did add comments below about documenting a bit what what did get re-run, and wasn't sure if there was some external guidelines of scoping out how much should be re-run).

### Paper Comments

My main comment of the paper, is that I think the paper could be a bit clearer about what it tried to reproduce, and in what ways this was / was not successful.

- For example, I think it would be easier to follow if there was a more clear & explicit note about retrieving the data, in the main text, rather than a link to the acknowledgements. Reading the paper, by the end I could piece together that some of the data was not retrieved, but I think this should be just directly stated directly and early, as we "we attempted to access the code, and found X, and for the data, were able to access Y".

- It is a bit unclear how successful the replication attempt was. In section 1.2, you say "reproducibility of the results was harder and only partly successful". I'm slightly unclear, even after reading the full paper, whether this refers specifically to not finding all of the data to re-run things completely, and/or if some things couldn't be run for other reasons. I would suggest being more explicit and separating / noting success or lack thereof due to data, and success or lack thereof due to software issues. The end of part 3 seems to allude to the notion that everything you tried to run did work in the end. If so, I think the last sentence of part 3 could clarify whether this relates to re-running _all_ key findings, or _selected_ key findings (from the original paper) as it's unclear how much was attempted (see next point).

- Relatedly, how much of the paper was attempted to be replicated? The paper reports and recreates 4 figure sub-panels, as a sub-selection of the original paper. It is unclear if these are selected as key elements of the original, and/or if they are the only ones able to be re-created with limited data, and/or if anything was attempted to be re-run and failed for software reasons. I'm not sure if there is a recommended threshold for how much should be attempted, but I would suggest that perhaps at least a brief description of how many figures were attempted to be re-run, and why these specific figures were included in the replication paper would be useful. One possibility might be to state more specifically what the key findings are of the original paper that you will directly reproduce here (including if this is a sub-selection). Maybe this could be something like "from the original paper, we selected key findings X, Y, Z, as demonstrated in original figures X1, X2, X3, and report on those figures here". If more was run, that is not reported in the ReScience paper, these extra figures could perhaps be added and noted in the repository?

- There are a couple parts that redirect to other sections, that I don't quite follow. In the second paragraph of part 3, it states finding scripts was partly successful "(see above)", but the above section seems to indicate you found all the scripts? I think maybe you are alluding to some difficulties finding parameter files (this also refers the reader to the acknowledgments). I think this section should be clearer on stating directly what was found, and what was not  (without referencing other sections). This description could / should perhaps be consolidated and moved into section 2.

- Figure 2 is included, I believe, to demonstrate that `rmouse` was able to be re-run? I think this makes sense, but that the figure caption (or perhaps somewhere in text) could acknowledge and describe this slightly. As of right now, there is no description to really understand how this figure was created or what the panels show. Perhaps the main text could briefly describe something like "to investigate the re-use of the rmouse tool, X data was used to generate sample plots, briefly described as Y". This could also / instead be described in the repository, but I didn't find any reference to making this figure in the repository.

- The guidelines for the challenge recommend saying whether "someone else would be able to do the required updates". This is something you could perhaps note in your final comments in the paper.

Small / Stylistic Notes:

- there is an opened but not closed parenthesis on line 5 of page 3
- page four, bullet point labelled '1', "two dozens" should be "two dozen"
- in the figure 2 caption. I'm not sure I understand the 'working style' part  - I think maybe those words could be dropped?
- figure 3, legend for D should reference that this figure is 9A from the original paper

### Code Notes

I do find the repository a little hard to understand what everything is. Could the layout of the files be described a bit more?

For example:

- in `scripts_paper` it might be easier if `Instructions_for_figures.md` was called `README.md`, so it would render in this folder
	- The first bullet point cuts off mid sentence
	- I'm not sure I really follow these instructions. They seem to describe running things in `scripts_data`. It's unclear what `reservoir` is.

- in `scripts_data`, I think there needs to be some kind of description of these files. What are the folders? Files for each subject?
	- Right now I don't know how to navigate these files. Indicating which are parameter files, if and where some kind of master scripts are, and what to focus on would be helpful

- Based on the guidelines, I think the repo is supposed to have a copy of the original code & of the updated version?
	- I think the repo only contains the updated version? Does the original commit in the repo reflect the original, unedited code?
	- Perhaps note on if all of this code has been updated and if / how to see the original version would be useful
