### 2020-05-26 #30 @hhentschke
@TomDonoghue Thank you very much for the very thorough review! I'll answer your points in detail once your review-related comments have been commented upon. Just a quick explanation of my take on the challenge. My understanding, garnered from the description of the challenge (https://rescience.github.io/ten-years/), was that a depiction of the journey was more important than the destination. Given that unsuccessful attempts at reproduction were admissable, provided the story is interesting, made me believe that a proof of concept of repeatability plus a selected choice of actual figure reproductions would be adequate. Also, compared to a few early submissions I peeked at, which were quite short (2-4 pages), my manuscript is lengthy; showing reproductions of all nine figures of the orignal paper in the manuscript would certainly be off-putting for readers.
That being said, I agree that this goal should be communicated early on. Will do so.
