### 2020-05-26 #30 @TomDonoghue
Hey @hhentschke - I agree, and sorry, I didn't mean to imply that I think the paper should _have_ to include all of the original figures! Related to that point my main suggestion would be describe a bit more explicitly the plan and sub-selection chosen for replication. I also understand the call as more of a description of the replication attempt, and did scan and see at least one other submission that explicitly talked about replicating a subset of the paper. In terms of if there is a threshold for how much should be re-run, I'll leave that to the editors to say anything more about.
