### 2020-06-15 aa064d7
removes scripts_paper/reservoir and an unused template file
