+++ scripts_paper/reservoir/fcomodp01.m
@@ -1,180 +1,0 @@
-function fComodp01
-% makes a plot of frequency comodulograms (ctrl and drug in one separate figures; 
-% principal channel and exploring only)
-
-global DS AP
-
-% load data from original results file (slow) and save extracted data
-saveMd='original';  
-% load previously extracted data (fast)
-saveMd='extracted';  
-
-
-rv={'fComod'};
-rvi=1;
-      
-behav={'exploring'};
-locChan={'IN 11'};
-nLocChan=length(locChan);
-% axis limits
-xl=[0 100];
-
-
-% --- prepare graphics
-labelscale('fontSz',8,'scaleFac',.6,'lineW',.5,'markSz',8); 
-
-
-ornt='portrait';
-figdir='c:\projects\hippocampus\rmouse\paper_atropine\rawFig\';
-printas='-djpeg99';[];
-printas='-dpsc2';
-% printas=[];
-
-% ------ main loop: files
-cnt=0;
-for fi=1:2
-  % --- paths, channels
-  rmouse_ini;
-  cd([WP.rootPath '\beta3_wtko\wt0001_04708']);
-  switch fi
-    case 1
-      a001_r1;
-    case 2
-      a003_r1;    
-  end
-  AP_beta3_wtko;
-  rmouse_APcheck;
-  rawCh=rmouse_chan;
-  % locate desired channels within results structure
-  for i=1:nLocChan
-    % indices for intra-electrode CC (AP.LFPInd is not needed..)
-    intra_ix(i)=strmatch(locChan{i},AP.rawChAnNm(AP.LFPInd),'exact');
-    % indices for inter-electrode CC
-    inter_ix(i)=strmatch(locChan{i},DS.rawCh(AP.allLFPIdx,1),'exact');  
-    % principal channels should be clear
-  end
-  if isempty(strfind(DS.dpath,':')), DS.dpath=[WP.rootPath DS.dpath]; end
-  
-  % --- load 
-  switch saveMd
-    case 'original'
-      load([AP.resPath '\' AP.resFn],'r');
-      F=r(1).comF;
-      % --- all about time: axes
-      xax=F;
-      if ~isempty(xl)
-        cix=find(xax>=xl(1) & xax<=xl(2));
-        xax=xax(cix);
-      else
-        cix=1:length(xax);
-      end
-      
-      % --- extract 
-      % determine index to results obtained with specified behavior in r
-      % note that intersect will sort alphabetically, which we dont want)
-      [nada,rix]=intersect({r(:).segmentType},behav);
-      c=[];
-      for bi=sort(rix)
-        % structure of c: freq|channel|behav
-        eval(['tmpc=r(bi).' rv{rvi} ';']);
-        switch rv{rvi}
-          case {'fComod'}
-            c=tmpc(:,:,intra_ix);
-        end
-      end
-      save([figdir DS.abfFn '_tmpDat'],'xax','cix','F','c');
-      clear tmpc r
-      
-    case 'extracted'
-      load([figdir DS.abfFn '_tmpDat']);
-  end      
-  
-  % --- plot
-  
-  % assume that bulk of correlation coeffs remain in that interval
-  ccLim=[-.20 .35];
-  nCol=255;
-
-  for loci=1:nLocChan
-    for bi=1:length(behav)
-      cnt=cnt+1;
-      figName=[mfilename '_' DS.abfFn '_' behav{bi} '_' locChan{loci}];
-      figure(cnt), clf;
-      orient(ornt);
-      
-      if 0
-        % colormap(coma('turquoisamber','ncols',nCol));
-        % colormap(coma('blueblackorange','ncols',nCol));
-        % colormap(coma('mib','ncols',nCol));
-        cm=colormap;
-      else
-        cm=[];
[…]
