+++ scripts_paper/reservoir/a001_thgaeccdemoplot.m
@@ -1,228 +1,0 @@
-global AP
-
-% invoke the data set-generating routine
-dset_04708001;
-
-AP.job={...
-  'det_artifacts',...       % detect artifacts in neural signals & generate time stamp list
-  'gen_btsl',...            % generate behavioral time stamp list
-};
-
-% ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-%  LENGTH OF DATA; CHANNELS; FILE NAMES; PRINT ETC
-% ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-% time block(s) to be processed in s (from t=0=start of recording).
-% May be given as 'full'
-AP.rawExcerpt=              'full';[0 30];
-
-% the channels to be analyzed (must be a subset of DS.rawChan)
-% NOTE: specify full names as given in the abf file header in cell 
-% array of strings, like {'IN 0','IN 10'}
-AP.rawChAnNm=               {'IN 11'};
-
-% the 'principal' channel (serving as a reference for e.g. CC computations) 
-% Same rules as above apply
-AP.rawChPrincNm=             {'IN 11'};
-
-% determines how to deal with previously computed results of the 'seg_' jobs:
-% 'replace' means they will be discarded entirely and replaced by the results 
-%   of the current run. This should be the default.
-% 'append' means that the results of the current run will be appended to or 
-%   overwrite results of previous runs. 
-% *** see NOTE of AP.job above! ***
-AP.saveMode=                 'replace';'append';
-
-% path to results file(s) - either full path or relative path (in the latter case 
-% rmouse_ini.m must contain the proper root path; the concatenation of both must be
-% equivalent to the full path)
-AP.resPath=                  'd:\projects\madison\rmouse\paper_atropine\rawFig';
-AP.resPath=                  'd:\hh';
-
-% name of the results file(s) 
-AP.resFn=                    [DS.abfFn '_thgaeCCDemo'];
-
-% directory for intermediate results files - will be generated automatically 
-% if not existent. make sure there is enough disk space (at least ten times 
-% the size of one experimental file)
-AP.strmDir=                  [DS.dpath '\streams'];
-
-% name of the log file (which collects information about each run of rmouse.m
-% on the data)
-AP.logFn=                    [DS.abfFn '_rubble.txt'];
-
-% save plot(s) to file in what format? '-dpsc2' (color postscript level 2) is
-% strongly recommended for high-quality figures, whereas '-djpeg90' (jpeg with 
-% quality level of 90) is more efficient for everyday use. Multiple formats can 
-% be specified; in this case they must be listed in a cell array, like 
-% {'-dpsc2','-djpeg90'}. Set to [] if plots shall not be saved.
-AP.printas=                  {'-dpsc2'};
-
-% ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-%        BEHAVIOR
-% ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-% data for behavioral scoring: file name without extension 
-% two data formats are supported:
-% 1. abf, in which case the file must contain trigger pulses as described below
-% 2. txt, a text file which must contain in its first column time stamps (in ms) 
-%    and in its second column a number corresponding to the trigger level (listed
-%    in column 2 of AP.bScoreFn below). The time stamps need not be sorted.
-% NOTES
-% a) If two files with identical names but different extensions are found the 
-%    txt file is used
-% b) In cases where behavioral scoring data was recorded in parallel with and 
[…]
