### 2020-06-15 bb92a8a
overhauls readme docs
