+++ paper/content.tex
@@ -1,5 +1,7 @@
+\section{Abstract}
+This article describes an attempt to reproduce key findings of a study of murine hippocampal oscillations published in 2007 \cite{hentschke_muscarinic_2007}. Being a submission to the 'Ten Years Reproducibility Challenge', it focuses on the process of reviving the Matlab code underlying the analyses, particularly the neuronal signal processing toolbox created for the study and its follow-ups.
+
 \section{Introduction}
-This paper describes an attempt to reproduce analyses and key findings in \cite{hentschke_muscarinic_2007}.
 \subsection{Neuroscientific background}
 \label{subsec:background}
 Neuronal activity in mammalian central nervous systems is often oscillatory. Intensity, frequency, and other characteristics of the oscillations are closely associated with behavior and arousal state. As this association suggests a functional role of oscillations, they have been studied extensively. One of the most conspicuous cases of behavior-dependent oscillatory neuronal activity occurs in hippocampus, a brain region pivotal for navigation and memory formation: prominent rhythms exist in the theta (4-12 Hz) and gamma (30-90 Hz) bands. 
@@ -7,12 +9,16 @@ Interestingly, the oscillations are coupled in a nonlinear fashion: the magnitud
 
 A large body of research has shown that both theta and gamma rhythms depend on the neuromodulator acetylcholine, but before our study \cite{hentschke_muscarinic_2007} it had not been clear whether and how their PAC would change with cholinergic signalling. As a first step towards shedding light on this, we performed experiments on mice with extracellular multielectrode arrays implanted into area CA1 of hippocampus. Each extracellular electrode picked up local field potentials (LFPs) -- tiny, local, time-dependent changes in the electrical potential relative to a distant reference point (Fig.~\ref{fig:hip_oscill}B). The mice were behaviorally scored and the LFPs recorded for periods of ca.\ 90 min, including administration of a blocker of muscarinic cholinergic receptors 30 min into the experiments. 
 
-The code presented here consists of two major parts. One is a toolbox of bespoke data crunching routines which performed classical time and frequency domain analyses. These were geared towards revealing spatial and temporal relations between signals from the different recording electrodes (crosscorrelations, coherence, etc.), with a focus on within-site phase-amplitude coupling. The 'main' function of this toolbox -- termed 'rmouse' -- integrated behavioral scoring data, partitioned the computational results according to the animals' behavior, and created 'working-style' figures for each recording (Fig.~\ref{fig:rmouse_workingstyle}). Also included were numerous post-processing routines with specific jobs like aggregation of data.
+\subsection{Code base and scope of reproduction}
+The code presented here consists of two major parts. One is a toolbox of bespoke data crunching routines which performed classical time and frequency domain analyses. These were geared towards revealing spatial and temporal relations between signals from the different recording electrodes (crosscorrelations, coherence, etc.), with a focus on within-site phase-amplitude coupling. The 'main' function of this toolbox -- termed 'rmouse' -- integrated behavioral scoring data, partitioned the computational results according to the animals' behavior, and created summary figures for each recording (Fig.~\ref{fig:rmouse_workingstyle}). Also included were numerous post-processing routines with specific jobs like aggregation of data.
 
 The other part of the code consisted of project-specific routines for statistical analyses of the data, and for the generation of publication-style plots.
 
+The major focus here is on getting the toolbox and associated post-processing code to run with a Matlab version as recent as possible. Moreover, I attempted to reproduce those figures of the original study which conveyed the key finding of the paper, namely a reduction of PAC in the presence of muscarinic blockers (Fig.~\ref{fig:repro}).
+
 \subsection{Expectations and results in a nutshell}
[…]
