### 2020-06-15 #30 @hhentschke
Hi @TomDonoghue, I decided to move forward and work on your comments, irrespective of the issue of completeness. You were quite right with the points you mentioned, and I changed the manuscript in the following ways:
* It is now clearly stated in the main text that only a subset of the experimental data could be retrieved, but that otherwise the reproduction was successful (in 'Expectations and results in a nutshell', also also at the end of section 4 (formerly 3)). I also cut down the verbose acknowledgements.

- There are now clear statements about the exact scope of the reproduction: in the (new) Abstract, and in the last paragraph of the new section 'Code base and scope of reproduction'. The exact figure panels from the paper I had in mind for reproduction are listed in the legend of Figure 3.

- It is stated that all other figure panels featuring the horizontal bar plots could be reproduced (end of section 4).

- The multiple and partly circular redirections to the various sections of the manuscript are gone; I have tried to be as explicit as possible.

- The purpose, underlying data and details of Figure 2 are given in the caption, and the term 'working-style' is gone.

- The 'Conclusion and personal evaluation' has been extended by a comment on the feasibility of the required changes to the code by anyone else.

- The small stylistic glitches have been taken care of.

As to the code and repository, I have reorganized, corrected and extended the documentation of the repository such that hopefully the structure of the code is much clearer now. In detail:

- In scripts_paper, Instructions_for_figures.md has been renamed Readme.md, as you suggested, and the content split up and clarified
- There is a new readme for scripts_data
- Subfolder 'reservoir' is gone as it wasn't very helpful
- the basic 'master script' is explicitly mentioned

The repository does in fact contain the code in its original, unchanged version, including a commit message explaining its state. See the initial commit from April 3rd this year. Including the latest commit, there is a total of 15 commits (https://github.com/hhentschke/ReScience_10yrReproChallenge/commits/master), documenting the progression of changes.

Once again, thanks for your efforts! I'm curious to hear what the editors say.
