### 2020-08-11 #30 @TomDonoghue
Hey @hhentschke (and cc: @oliviaguest):

First of all, I would like to apologize for the large delay in replying here. It seems I missed the original notification, and happened to notice this when scrolling through open issues, so my bad and my apologies for that.

I have revisited the project, and checked the updated repository and the updated paper. I am happy with the updates to the paper, and I find the organization of the updated version to be much easier to follow. I think the updates to the repository also help to organize and describe the files. When I re-read the paper now, I don't have any other notable comments or concerns about this paper.

In terms of my original questions to the editors - I think the updated paper is quite clear on replicating the usage of the a processing toolbox and a key set of results from the original paper, and I do not have any worries about the extent of the replication. As far as I'm aware, reviewing and/or re-running the code is not strictly required. Overall, I don't feel there is any outstanding issues that necessarily require anything more here.

All in all, I consider this paper is an interesting and useful submission for the 10 Years Challenge, as I understand the call, and in my capacity as reviewer, I would be happy to say this submission looks good to me to be accepted for that.

Thank you for the work on the updates @hhentschke - I enjoyed reading and reviewing this paper, congrats on the interesting submission!

@oliviaguest - please let me know if there is anything else you need from me for this submission!
