### 2020-09-09 #30 @oliviaguest
@hhentschke Yes, can you please fill in the missing details in the https://github.com/hhentschke/ReScience_10yrReproChallenge/blob/master/paper/metadata.yaml file, please? Set acceptance date as today. 🎉

This is relevant info: https://github.com/ReScience/ReScience/issues/48#issuecomment-689405707. And my ORCiD is: https://orcid.org/0000-0002-1891-0972 — let me know if anything is not clear. ☺️
