+++ paper/metadata.yaml
@@ -32,8 +32,8 @@ keywords: rescience c, Matlab, hippocampus, cross-frequency coupling, phase-ampl
 # Software Heritage.
 #   see https://guides.github.com/activities/citable-code/
 code:
-  - url: 
-  - doi: 
+  - url: https://github.com/hhentschke/ReScience_10yrReproChallenge/tree/v.1.0.0
+  - doi: 10.5281/zenodo.4021389
   - swh: 
 
 # Data URL and DOI (optional if no data)
@@ -75,13 +75,13 @@ language: Matlab
 
 # For example, the URL of the GitHub issue where review actually occured
 review: 
-  - url: 
+  - url: https://github.com/ReScience/submissions/issues/30
 
 contributors:
-  - name:
-    orcid: 
+  - name: Olivia Guest
+    orcid: 0000-0002-1891-0972
     role: editor
-  - name:
+  - name: Tom Donoghue
     orcid:
     role: reviewer
   - name:
@@ -90,9 +90,9 @@ contributors:
 
 # This information will be provided by the editor
 dates:
-  - received:  November 1, 2018
-  - accepted:
-  - published: 
+  - received:  April 21, 2020
+  - accepted: September 9, 2020
+  - published: September 9, 2020
 
 # This information will be provided by the editor
 article:
@@ -104,5 +104,5 @@ article:
 journal:
   - name:   "ReScience C"
   - issn:   2430-3658
-  - volume: 4
-  - issue:  1
+  - volume: 6
+  - issue:  3
