### 2020-09-09 #30 @hhentschke
Yay! Thanks a lot for your efforts! I left the last but one block ("article") in the metadata.yaml open because I'm not sure whether that final step of assigning article number, DOI and pdf url is up to you (including re-making the file) or me. Will do so if you tell me to.
