### 2020-09-10 #30 @oliviaguest
@rougier is this all done now? How do we update the website?

@hhentschke please merge https://github.com/hhentschke/ReScience_10yrReproChallenge/pull/1 😊 when you can!
